Average tempo-tapping over multiple inputs
This change adds a queue of tap-timings and a settings to control the length of the queue.
The BPM value is calculated by taking the average of the times in the queue.
This results in greater precision, by allowing some off-timed presses under the assumption that the distribution is sort
of normally distributed around the desired tempo.
Setting 2 taps (= 1 value in the queue) makes the behavior identical to the previous version.